requirements.txt
matplotlib==3.7.2
network==0.1
numpy==1.24.3
opencv_python==4.8.0.76
openslide_python==1.3.0
Pillow==10.0.0
scikit_image==0.19.3
SimpleITK==2.2.1
torch==2.0.0
